fix: 7 verified bugs from code review — async_write errors, timer pending leak, UB shifts, atomic phase
BUG: connection_impl.hpp async_write handler silently ignored errors
(used `auto, auto` discarding error_code). Write failures now detected,
queue drained, and connection closed via close_internal(connection_aborted).

BUG: connection_impl.hpp receive handler lacked try/catch around
m_connection_handler dispatch. Exception now caught, logged, and
connection closed instead of crashing the io_context event loop.

BUG: connection_impl.hpp RECV length parsing had UB — uint8_t promoted
to signed int before << 24, shifting into sign bit. Added static_cast<uint32_t>.

BUG: LLP/utils.hpp bytes2uint() had same UB — uint8_t << 24 without cast.
Added static_cast<uint32_t> on all shift operands.

BUG: worker_manager.cpp forced retry timer error path returned without
clearing m_forced_retry_timer_pending, permanently blocking future retries
after timer cancellation.

BUG: worker_manager.hpp RecoveryContext::phase and degraded_signal were
plain types despite cross-thread access from signal handler path.
Made atomic with relaxed/release ordering. Updated all read/write sites.

CLEANUP: chacha20_wrapper.cpp is_available() allocated EVP_CIPHER_CTX
that was never used. Simplified to just check EVP_chacha20_poly1305().

NOT A BUG: session_manager.cpp append_uint32_le() is correct LE encoding.

Agent-Logs-Url: https://github.com/NamecoinGithub/NexusMiner/sessions/dd9c74dd-9e76-437f-93d0-2074e1fe2109

+++ src/LLP/utils.hpp
@@ -60,7 +60,7 @@ inline uint32_t bytes2uint(std::vector<uint8_t> const& BYTES, int nOffset = 0)
 	if (BYTES.size() < nOffset + 4)
 		return 0;
 
-	return (BYTES[0 + nOffset] << 24) + (BYTES[1 + nOffset] << 16) + (BYTES[2 + nOffset] << 8) + BYTES[3 + nOffset];
+	return (static_cast<uint32_t>(BYTES[0 + nOffset]) << 24) + (static_cast<uint32_t>(BYTES[1 + nOffset]) << 16) + (static_cast<uint32_t>(BYTES[2 + nOffset]) << 8) + static_cast<uint32_t>(BYTES[3 + nOffset]);
 }
 
 /** Convert a 64 bit Unsigned Integer to Byte Vector using Bitwise Shifts. **/
